requirements.txt
uvicorn
fastapi
cohere
python-dotenv
chromadb
langchain
langchain_chroma
langchain_cohere
